requirements.txt
flask==2.3.3
flask-socketio==5.3.4
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.2
# Removed all Hugging Face dependencies
torch==2.0.1
faiss-cpu==1.7.4
google-generativeai==0.3.2
weaviate-client==3.18.0
pymongo==4.5.0
motor==3.3.1
sqlalchemy==2.0.23
pandas==2.0.3
pydantic==2.0.3
python-dateutil==2.8.2
pytz==2023.3
# No longer need accelerate since we removed the local transformer models

# Voice and facial recognition dependencies
opencv-python==4.8.0.74
SpeechRecognition==3.10.0
gTTS==2.3.2
pydub==0.25.1
bcrypt==4.0.1
PyJWT==2.8.0
python-socketio==5.8.0
eventlet==0.33.3

# Added for incremental training scheduler
schedule==1.2.0